+++ appTranslation.py
@@ -31,9 +31,10 @@ languages_dict = {
     'fr': 'French',
     'hu': 'Hungarian',
     'it': 'Italian',
+    'pt_BR': 'Brazilian Portuguese',
     'ro': 'Romanian',
     'ru': 'Russian',
-    'pt_BR': 'Brazilian Portuguese',
+    'tr': 'Turkish',
 }
 
 translations = {}